add entropy calc back in
went through all the effort of accurately calculating entropy when you
disallow consecutive spaces and... it makes no difference (except in the
tiniest of cases, like three character passwords with only a few
characters enabled). Which is unsurprising because it amounts to
sometimes not being able to use *one* character, and normally you have
to enable/disable *many* characters to see any entropy difference.